+++ Nexmo.Api/Voice/Call.cs
@@ -9,14 +9,83 @@ namespace Nexmo.Api.Voice
     {
         public class Endpoint
         {
+            /// <summary>
+            /// One of the following: phone, websocket, sip
+            /// </summary>
             public string type { get; set; }
+
+            // phone:
+
+            /// <summary>
+            /// The phone number to connect to in E.164 format.
+            /// </summary>
             public string number { get; set; }
+            /// <summary>
+            /// Set the digits that are sent to the user as soon as the Call is answered. The * and # digits are respected. You create pauses using p. Each pause is 500ms.
+            /// </summary>
+            public string dtmfAnswer { get; set; }
+
+            // websocket/sip:
+
+            /// <summary>
+            /// The URI to the websocket you are streaming to.
+            /// OR
+            /// The SIP URI to the endpoint you are connecting to in the format sip:rebekka@sip.example.com.
+            /// </summary>
+            public string uri { get; set; }
+
+            // websocket:
+
+            /// <summary>
+            /// The internet media type for the audio you are streaming.Possible values are: audio/l16; rate=16000
+            /// </summary>
+            [JsonProperty("content-type")]
+            public string contentType { get; set; }
+
+            /// <summary>
+            /// A JSON object containing any metadata you want.
+            /// </summary>
+            public string headers { get; set; }
         }
+
         public class CallCommand
         {
+            /// <summary>
+            /// The single or mixed collection of endpoint types you connected to. Possible values.
+            /// </summary>
             public Endpoint[] to { get; set; }
+            /// <summary>
+            /// The endpoint you are calling from. Possible value are the same as to.
+            /// </summary>
             public Endpoint from { get; set; }
+            /// <summary>
+            /// The webhook endpoint where you provide the Nexmo Call Control Object that governs this call. As soon as your user answers a call, Platform requests this NCCO from answer_url. Use answer_method to manage the HTTP method.
+            /// </summary>
             public string[] answer_url { get; set; }
+            /// <summary>
+            /// Optional. The HTTP method used to send event information to answer_url. The default value is GET.
+            /// </summary>
+            public string answer_method { get; set; }
+            /// <summary>
+            /// Optional. Platform sends event information asynchronously to this endpoint when status changes. For more information about the values sent, see callback.
+            /// </summary>
+            public string event_url { get; set; }
+            /// <summary>
+            /// Optional. The HTTP method used to send event information to event_url. The default value is POST.
+            /// </summary>
+            public string event_method { get; set; }
+            /// <summary>
+            /// Optional. Configure the behavior when Nexmo detects that a destination is an answerphone.
+            /// </summary>
+            public string machine_detection { get; set; }
+            /// <summary>
+            /// Optional. Set the number of seconds that elapse before Nexmo hangs up after the call state changes to in_progress. The default value is 7200, two hours. This is also the maximum value.
+            /// </summary>
+            public decimal length_timer { get; set; }
+            /// <summary>
+            /// Optional. Set the number of seconds that elapse before Nexmo hangs up after the call state changes to 'ringing'. The default value is 60, the maximum value is 120.
+            /// </summary>
+            public decimal ringing_timer { get; set; }
         }
 
         public class CallCommandResponse
@@ -35,9 +104,18 @@ namespace Nexmo.Api.Voice
         {
             /// <summary>
             /// Use one of the following options to modify the call:
-            ///  hangup - end this Call.
+            /// hangup    Terminates this call leg.
+            /// mute      Mutes this call leg.
+            /// unmute    Unmutes this call leg.
+            /// earmuff   Prevents the recipient of this call leg from hearing other parts of the conversation.
+            /// unearmuff Removes the earmuff effect from this call leg.
+            /// transfer  Transfers this call leg to another NCCO, as specified by the destination parameter.
             /// </summary>
             public string action { get; set; }
+            /// <summary>
+            /// Optional. A JSON object pointing to a replacement NCCO, when action is transfer.
+            /// </summary>
+            public string destination { get; set; }
         }
 
         public class SearchFilter
@@ -102,6 +180,22 @@ namespace Nexmo.Api.Voice
             /// </summary>
             public Endpoint from { get; set; }
             /// <summary>
+            /// Possible values are outbound or inbound.
+            /// </summary>
+            public string direction { get; set; }
+            /// <summary>
+            /// The URL to download a call or conversation recording from.
+            /// </summary>
+            public string recording_url { get; set; }
+            /// <summary>
+            /// The time the Call started
+            /// </summary>
+            public DateTime? start_time { get; set; }
+            /// <summary>
+            /// The Mobile Country Code Mobile Network Code (MCCMNC) for the carrier network used to make this Call.
+            /// </summary>
+            public string network { get; set; }
+            /// <summary>
             /// The status of the Call. Possible values are:
             ///  started - Platform has stared the Call.
             ///  ringing - the user's handset is ringing.
@@ -111,10 +205,6 @@ namespace Nexmo.Api.Voice
             ///  completed - Platform has terminated this Call.
             /// </summary>
             public string status { get; set; }
-            /// <summary>
-            /// Possible values are outbound or inbound.
-            /// </summary>
-            public string direction { get; set; }
             /// <summary>
             /// The price per minute for this Call.
             /// </summary>
@@ -128,17 +218,9 @@ namespace Nexmo.Api.Voice
             /// </summary>
             public int duration { get; set; }
             /// <summary>
-            /// The time the Call started
-            /// </summary>
-            public DateTime start_time { get; set; }
-            /// <summary>
             /// The time the Call ended
             /// </summary>
-            public DateTime end_time { get; set; }
-            /// <summary>
-            /// The Mobile Country Code Mobile Network Code (MCCMNC) for the carrier network used to make this Call.
-            /// </summary>
-            public string network { get; set; }
+            public DateTime? end_time { get; set; }
         }
 
         public class CallList